+++ src/x12sdk/v5010/x12_271_005010X279A1/parsing.py
@@ -118,6 +118,7 @@ def set_information_source_name_loop(
     """
     if context.loop_name == TransactionLoops.INFORMATION_SOURCE:
         context.loop_container[TransactionLoops.INFORMATION_SOURCE_NAME] = {
+            "prv_segment": [],
             "per_segment": [],
             "aaa_segment": [],
         }